GOCBC-1769: Change to Authenticator interface
Motivation
----------
We need a way to do mtls rotation and also to support JWT auth.
To achieve this we need to refactor how auth is done.
This is technically a breaking change but we are not aware of any
users other than gocb.

Changes
-------
Remove username, password, and certificate.
Create Authenticator interface and certificate and basic authenticators.
Expose a ReconfigureAuthenticator function on routing client.